+++ src/parsers/parseCronDayOfWeek.ts
@@ -1,4 +1,9 @@
 import { CronDayOfWeek } from "../types/CronDayOfWeek";
+import { CronLast } from "../types/CronLast";
+import { CronLastValue } from "../types/CronLastValue";
+import { CronSpecificDayOfWeek } from "../types/CronSpecificDayOfWeek";
+import { CronStartOrBlank } from "../types/CronStartOrBlank";
+import { parseCronDayOfWeekValue } from "./parseCronDayOfWeekValue";
 import { parseCronLast } from "./parseCronLast";
 import { parseCronLastValue } from "./parseCronLastValue";
 import { parseCronPart } from "./parseCronPart";
@@ -9,9 +14,14 @@ import { parseCronStartOrBlank } from "./parseCronStartOrBlank";
  * Parses given source CronDayOfWeek to string.
  * @param source Source CronDayOfWeek.
  */
-export const parseCronDayOfWeek = (source: CronDayOfWeek): string =>
-	parseCronLast(source) ??
-	parseCronStartOrBlank(source) ??
-	parseCronLastValue(source) ??
-	parseCronSpecificDayOfWeek(source) ??
-	parseCronPart(source);
+export const parseCronDayOfWeek = (source: CronDayOfWeek) =>
+	parseCronLast(source as CronLast) ??
+	parseCronStartOrBlank(source as CronStartOrBlank) ??
+	parseCronLastValue(source as CronLastValue) ??
+	parseCronSpecificDayOfWeek(source as CronSpecificDayOfWeek) ??
+	parseCronPart(parseCronDayOfWeekValue)(
+		source as Exclude<
+			CronDayOfWeek,
+			CronLast | CronStartOrBlank | CronLastValue | CronSpecificDayOfWeek
+		>
+	);