+++ src/parsers/parseCronLast.ts
@@ -1,8 +1,9 @@
+import { CronLast } from "../types/CronLast";
 import { isCronLast } from "../validations/isCronLast";
 
 /**
  * Parse given source string to ParseLast.
  * @param source Source string to be parsed.
  */
-export const parseCronLast = (source: unknown) =>
+export const parseCronLast = (source: CronLast | string) =>
 	isCronLast(source) ? source : undefined;