+++ src/parsers/parseCronList.ts
@@ -1,6 +1,9 @@
-import { arrayMap } from "@vangware/utils";
+import { arrayMap, isUndefined } from "@vangware/utils";
 import { CRON_LIST_SEPARATOR } from "../constants";
-import { CronListItem } from "../types/CronListItem";
+import { CronList } from "../types/CronList";
+import { CronRange } from "../types/CronRange";
+import { CronSteps } from "../types/CronSteps";
+import { CronValueParser } from "../types/CronValueParser";
 import { isCronList } from "../validations/isCronList";
 import { parseCronRange } from "./parseCronRange";
 import { parseCronSteps } from "./parseCronSteps";
@@ -10,10 +13,19 @@ import { parseCronSteps } from "./parseCronSteps";
  * array (or undefined if is invalid).
  * @param source Source string to be parsed.
  */
-export const parseCronList = <Value>(source: unknown) => {
-	const list = arrayMap<CronListItem<Value>, string>(
-		value => parseCronSteps(value) ?? parseCronRange(value) ?? `${value}`
-	)(isCronList<Value>(source) ? source : []).join(CRON_LIST_SEPARATOR);
+export const parseCronList = <Value>(parser: CronValueParser<Value>) =>
+	/**
+	 * @param source Source string to be parsed.
+	 */
+	(source: CronList<Value>) => {
+		const list = arrayMap(
+			value =>
+				parseCronSteps(parser)(value as CronSteps<Value>) ??
+				parseCronRange(parser)(value as CronRange<Value>) ??
+				parser(value as Value)
+		)(isCronList<Value>(source) ? source : []);
 
-	return list.length === 0 ? undefined : list;
-};
+		return list.length === 0 || list.some(isUndefined)
+			? undefined
+			: list.join(CRON_LIST_SEPARATOR);
+	};