+++ src/parsers/parseCronMonth.ts
@@ -1,8 +1,8 @@
-import { CronMonth } from "../types/CronMonth";
+import { parseCronMonthValue } from "./parseCronMonthValue";
 import { parseCronPart } from "./parseCronPart";
 
 /**
  * Parses given source CronMonth to string.
  * @param source Source CronMonth.
  */
-export const parseCronMonth = (parse: CronMonth) => parseCronPart(parse);
+export const parseCronMonth = parseCronPart(parseCronMonthValue);