+++ src/parsers/parseCronStartOrBlank.ts
@@ -1,8 +1,9 @@
+import { CronStartOrBlank } from "../types/CronStartOrBlank";
 import { isCronStartOrBlank } from "../validations/isCronStartOrBlank";
 
 /**
  * Parse given source string to CronStartOrBlank.
  * @param source Source string to be parsed.
  */
-export const parseCronStartOrBlank = (source: unknown) =>
+export const parseCronStartOrBlank = (source: CronStartOrBlank | string) =>
 	isCronStartOrBlank(source) ? source : undefined;