+++ src/validations/isStringNearestDayOfMonth.ts
@@ -1,6 +1,8 @@
+import { isString } from "@vangware/utils";
+
 /**
  * Check if given value is "{value}W".
  * @param value Value to check.
  */
-export const isStringNearestDayOfMonth = (value: string) =>
-	/^(?:[1-9]|[12][0-9]|3[01])W$/iu.test(value);
+export const isStringNearestDayOfMonth = (value: unknown): value is string =>
+	isString(value) && /^(?:[1-9]|[12][0-9]|3[01])W$/iu.test(value);