+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 This project uses [Gitmoji](https://gitmoji.carloscuesta.me/)
 
+## 2.0.0
+
+- :building_construction: had to change cron parsers to return undefined when invalid to be more consistent.
+- :white_check_mark: updated tests (have to add more later).
+
 ## 1.0.2
 
 - :bug: fix output issue for esm (too modern for everyone).